// src/types/airtableDataset/FieldSchema.ts
export type GenericPreSanitizationValueInputProps<Fields extends GenericRecord = GenericRecord> = {
  airtableSchema: AirtableSchema;
  rawAirtableDataset: RawAirtableDataset;
  tableName: string;
  tableSchema: TableSchema;
  fieldSchemaName: string;
  fieldName: string;
  fieldSchema: FieldSchema;
  rawFieldValue: any;
  sanitizedFieldValue: any;
  rawRecord: RawAirtableRecord<Fields>;
  preferredLocalesOrLanguages: string[];
};

/**
 * Input props received by all "post-sanitization" functions that are called at the beginning of the sanitization process.
 */
export type GenericPostSanitizationValueInputProps<Fields extends GenericRecord = GenericRecord> = GenericPreSanitizationValueInputProps & {
  sanitizedRecord: AirtableRecord<Fields>;
};

/**
 * Input props received by all "post-consolidation" functions that are called at the end of the consolidation process.
 */
export type GenericPostConsolidationTransformationValueInputProps<Fields extends GenericRecord = GenericRecord> = {
  airtableSchema: AirtableSchema;
  sanitizedDataset: SanitizedAirtableDataset;
  tableName: string;
  tableSchema: TableSchema;
  fieldSchemaName: string;
  fieldName: string;
  fieldSchema: FieldSchema;
  sanitizedFieldValue: any;
  sanitizedRecord: AirtableRecord<Fields>;
};